+++ NEWS.md
@@ -1,5 +1,7 @@
 # valr (development version)
 
+* `read_bed` and related functions now automatically calculate the fields. Use of `n_fields` was deprecated.
+
 # valr 0.6.8
 
 * `bed_closest()` now reports all x intervals, even when there are no closest y intervals (e.g. when there is no matching chromosome in y intervals). These intervals are returned populated with `NA` for `.overlap`, `.dist` and y interval locations.